backend-fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "pytz>=2025.2",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
    "web3>=6.0.0",
    "eth-account>=0.5.9",
    "py-solc-x>=2.0.0",
]
